+++ Revit_Core_Engine/Convert/MEP/ToRevit/PipeSegment.cs
@@ -47,7 +47,7 @@ namespace BH.Revit.Engine.Core
         [Input("settings", "Revit adapter settings to be used while performing the convert.")]
         [Input("refObjects", "Optional, a collection of objects already processed in the current adapter action, stored to avoid processing the same object more than once.")]
         [Output("pipeSegment", "Revit PipeSegment resulting from converting the input BH.oM.MEP.System.MaterialFragments.PipeMaterial.")]
-        public static PipeSegment ToRevitPipeSegement(this PipeMaterial pipeMaterial, Document document, RevitSettings settings = null, Dictionary<Guid, List<int>> refObjects = null)
+        public static PipeSegment ToRevitPipeSegement(this PipeMaterial pipeMaterial, Document document, RevitSettings settings = null, Dictionary<Guid, List<long>> refObjects = null)
         {
             if (document == null || pipeMaterial == null)
                 return null;
