+++ Revit_Core_Engine/Convert/Physical/ToRevit/Material.cs
@@ -43,7 +43,7 @@ namespace BH.Revit.Engine.Core
         [Input("settings", "Revit adapter settings to be used while performing the convert.")]
         [Input("refObjects", "Optional, a collection of objects already processed in the current adapter action, stored to avoid processing the same object more than once.")]
         [Output("material", "Revit Material resulting from converting the input BH.oM.Physical.Materials.Material.")]
-        public static Material ToRevitMaterial(this BHP.Material material, Document document, RevitSettings settings = null, Dictionary<Guid, List<int>> refObjects = null)
+        public static Material ToRevitMaterial(this BHP.Material material, Document document, RevitSettings settings = null, Dictionary<Guid, List<long>> refObjects = null)
         {
             if (material == null)
                 return null;
